This commit was generated by cvs2svn to compensate for changes in r8271, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@8272 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/drive/riot2d.c
@@ -26,7 +26,7 @@
 
 struct drive_context_s;
 #define RIOT_SHARED_CODE
-#define RIOTCONTEXT	struct drive_context_s
+#define RIOTCONTEXT     struct drive_context_s
 
 #include "riotcore.h"
 
@@ -36,13 +36,13 @@ struct drive_context_s;
  * Renaming exported functions
  */
 
-#define myriot_init 	riot2_init
-#define myriot_signal	riot2_signal
-#define myriot_reset	riot2_reset
-#define myriot_store	riot2_store
-#define myriot_read	riot2_read
-#define myriot_peek	riot2_peek
-#define myriot_set_flag	riot2_set_flag
+#define myriot_init     riot2_init
+#define myriot_signal   riot2_signal
+#define myriot_reset    riot2_reset
+#define myriot_store    riot2_store
+#define myriot_read     riot2_read
+#define myriot_peek     riot2_peek
+#define myriot_set_flag riot2_set_flag
 #define myriot_write_snapshot_module riot2_write_snapshot_module
 #define myriot_read_snapshot_module riot2_read_snapshot_module
 
@@ -60,12 +60,12 @@ struct drive_context_s;
 
 /*
 #define my_set_irq(fl, clk)                          \
-	do {                                         \
-	printf("set_int_d0(%d)\n",(fl));             \
+        do {                                         \
+        printf("set_int_d0(%d)\n",(fl));             \
         interrupt_set_irq(&(ctxptr->cpu.int_status), \
                           (ctxptr->riot2p.irq_type), \
                           (fl) ? IK_IRQ : 0, (clk))  \
-	; } while(0)
+        ; } while(0)
 */
 #define my_set_irq(fl, clk)                          \
         interrupt_set_irq(&(ctxptr->cpu.int_status), \
@@ -92,23 +92,23 @@ struct drive_context_s;
 
 
 /* renaming formerly global symbols */
-#define riotio		(ctxptr->riot2.riot_io)
-#define oldpa		(ctxptr->riot2.old_pa)
-#define oldpb		(ctxptr->riot2.old_pb)
-#define riot_log	(ctxptr->riot2.log)
-#define riot_alarm	(ctxptr->riot2.alarm)
-#define riot_read_clk	(ctxptr->riot2.read_clk)
+#define riotio          (ctxptr->riot2.riot_io)
+#define oldpa           (ctxptr->riot2.old_pa)
+#define oldpb           (ctxptr->riot2.old_pb)
+#define riot_log        (ctxptr->riot2.log)
+#define riot_alarm      (ctxptr->riot2.alarm)
+#define riot_read_clk   (ctxptr->riot2.read_clk)
 #define riot_read_offset (ctxptr->riot2.read_offset)
-#define riot_last_read	(ctxptr->riot2.last_read)
-#define edgectrl	(ctxptr->riot2.r_edgectrl)
-#define irqfl		(ctxptr->riot2.r_irqfl)
-#define irqline		(ctxptr->riot2.r_irqline)
-#define ti_write_clk	(ctxptr->riot2.r_write_clk)
-#define ti_N		(ctxptr->riot2.r_N)
-#define ti_divider	(ctxptr->riot2.r_divider)
-#define ti_irqen	(ctxptr->riot2.r_irqen)
-#define atn_active	(ctxptr->riot2p.r_atn_active)
-#define MYRIOT_NAME	(ctxptr->riot2.myname)
+#define riot_last_read  (ctxptr->riot2.last_read)
+#define edgectrl        (ctxptr->riot2.r_edgectrl)
+#define irqfl           (ctxptr->riot2.r_irqfl)
+#define irqline         (ctxptr->riot2.r_irqline)
+#define ti_write_clk    (ctxptr->riot2.r_write_clk)
+#define ti_N            (ctxptr->riot2.r_N)
+#define ti_divider      (ctxptr->riot2.r_divider)
+#define ti_irqen        (ctxptr->riot2.r_irqen)
+#define atn_active      (ctxptr->riot2p.r_atn_active)
+#define MYRIOT_NAME     (ctxptr->riot2.myname)
 
 void riot2_setup_context(drive_context_t *ctxptr)
 {
@@ -128,35 +128,35 @@ void riot2_setup_context(drive_context_t *ctxptr)
 _RIOT_FUNC void set_handshake(drive_context_t *ctxptr, BYTE pa)
 {
     /* IEEE handshake logic (named as in schematics):
-	Inputs: /ATN	= inverted IEEE atn (true = active)
-		ATNA	= pa bit 0
-		/DACO	= pa bit 1
-		RFDO	= pa bit 2
-	Output:	DACO	= /DACO & (ATN | ATNA) -> to IEEE via MC3446
-		RFDO	= (/ATN == ATNA) & RFDO -> to IEEE via MC3446
+        Inputs: /ATN    = inverted IEEE atn (true = active)
+                ATNA    = pa bit 0
+                /DACO   = pa bit 1
+                RFDO    = pa bit 2
+        Output: DACO    = /DACO & (ATN | ATNA) -> to IEEE via MC3446
+                RFDO    = (/ATN == ATNA) & RFDO -> to IEEE via MC3446
     */
     /* RFDO = (/ATN == ATNA) & RFDO */
     ctxptr->func.parallel_set_nrfd((char)(
-	!( ((atn_active ? 1 : 0) == (pa & 1)) && (pa & 4) ) 
-	));
+        !( ((atn_active ? 1 : 0) == (pa & 1)) && (pa & 4) )
+        ));
     /* DACO = /DACO & (ATNA | ATN) */
     ctxptr->func.parallel_set_ndac((char)(
-	!( (!(pa & 2)) && ((pa & 1) || (!atn_active)) )
-	));
+        !( (!(pa & 2)) && ((pa & 1) || (!atn_active)) )
+        ));
 }
 
 void drive_riot_set_atn(drive_context_t *ctxptr, int state)
 {
     if (DRIVE_IS_OLDTYPE(ctxptr->drive_ptr->type)) {
-	if (atn_active && !state) {
-	    riot2_signal(ctxptr, RIOT_SIG_PA7, RIOT_SIG_FALL);
-	} else
-	if (state && !atn_active) {
-	    riot2_signal(ctxptr, RIOT_SIG_PA7, RIOT_SIG_RISE);
-	}
+        if (atn_active && !state) {
+            riot2_signal(ctxptr, RIOT_SIG_PA7, RIOT_SIG_FALL);
+        } else
+        if (state && !atn_active) {
+            riot2_signal(ctxptr, RIOT_SIG_PA7, RIOT_SIG_RISE);
+        }
         atn_active = state;
-	riot1_set_pardata(ctxptr);
-	set_handshake(ctxptr, oldpa);
+        riot1_set_pardata(ctxptr);
+        set_handshake(ctxptr, oldpa);
     }
 }
 
@@ -164,8 +164,8 @@ _RIOT_FUNC void undump_pra(drive_context_t *ctxptr, BYTE byte)
 {
     /* bit 0 = atna */
     set_handshake(ctxptr, byte);
-    ctxptr->func.parallel_set_eoi(!(byte & 0x08));
-    ctxptr->func.parallel_set_dav(!(byte & 0x10));
+    ctxptr->func.parallel_set_eoi((BYTE)(!(byte & 0x08)));
+    ctxptr->func.parallel_set_dav((BYTE)(!(byte & 0x10)));
 }
 
 _RIOT_FUNC void store_pra(drive_context_t *ctxptr, BYTE byte)
@@ -176,8 +176,8 @@ _RIOT_FUNC void store_pra(drive_context_t *ctxptr, BYTE byte)
     /* bit 3 = eoio */
     /* bit 4 = davo */
     set_handshake(ctxptr, byte);  /* handle atna, nrfd, ndac */
-    ctxptr->func.parallel_set_eoi(!(byte & 0x08));
-    ctxptr->func.parallel_set_dav(!(byte & 0x10));
+    ctxptr->func.parallel_set_eoi((BYTE)(!(byte & 0x08)));
+    ctxptr->func.parallel_set_dav((BYTE)(!(byte & 0x10)));
 }
 
 _RIOT_FUNC void undump_prb(drive_context_t *ctxptr, BYTE byte)
@@ -190,7 +190,7 @@ _RIOT_FUNC void undump_prb(drive_context_t *ctxptr, BYTE byte)
     ctxptr->drive_ptr->led_status = (byte >> 4) & 0x03;
 
     if ((ctxptr->mynumber == 0) && (DRIVE_IS_DUAL(ctxptr->drive_ptr->type))) {
-	drive[1].led_status = ((byte & 8) ? 1 : 0) | ((byte & 32) ? 2 : 0);
+        drive[1].led_status = ((byte & 8) ? 1 : 0) | ((byte & 32) ? 2 : 0);
     }
 }
 
@@ -204,7 +204,7 @@ _RIOT_FUNC void store_prb(drive_context_t *ctxptr, BYTE byte)
     ctxptr->drive_ptr->led_status = (byte >> 4) & 0x03;
 
     if ((ctxptr->mynumber == 0) && (DRIVE_IS_DUAL(ctxptr->drive_ptr->type))) {
-	drive[1].led_status = ((byte & 8) ? 1 : 0) | ((byte & 32) ? 2 : 0);
+        drive[1].led_status = ((byte & 8) ? 1 : 0) | ((byte & 32) ? 2 : 0);
     }
 }
 
@@ -224,22 +224,27 @@ _RIOT_FUNC void riot_reset(drive_context_t *ctxptr)
 _RIOT_FUNC BYTE read_pra(drive_context_t *ctxptr)
 {
     BYTE byte = 0xff;
-    if (!parallel_atn) byte -= 0x80;
-    if (parallel_dav) byte -= 0x40;
-    if (parallel_eoi) byte -= 0x20;
+    if (!parallel_atn)
+        byte -= 0x80;
+    if (parallel_dav)
+        byte -= 0x40;
+    if (parallel_eoi)
+        byte -= 0x20;
     return (byte & ~riotio[1]) | (riotio[0] & riotio[1]);
 }
 
 _RIOT_FUNC BYTE read_prb(drive_context_t *ctxptr)
 {
     BYTE byte = 0xff;
-    if (parallel_nrfd) byte -= 0x80;
-    if (parallel_ndac) byte -= 0x40;
+    if (parallel_nrfd)
+        byte -= 0x80;
+    if (parallel_ndac)
+        byte -= 0x40;
 
     if (ctxptr->mynumber == 0)
-      byte -= 1;	/* device address bit 0 */
-    byte -= 2;		/* device address bit 1 */
-    byte -= 4;		/* device address bit 2 */
+        byte -= 1;        /* device address bit 0 */
+    byte -= 2;          /* device address bit 1 */
+    byte -= 4;          /* device address bit 2 */
 
     return (byte & ~riotio[3]) | (riotio[2] & riotio[3]);
 }
